Parameter extraction with Puppet 2.6
Increased backward compatibility with Puppet < 2.7, having no
instanciate method.

I use the presence of the instanciate method to use an alternate way of
getting hostclasses.

The classes are appended to the environment, so I use a Set of already
defined classes to append only the newly parsed one.

Tested with Puppet 2.6.2-5+squeeze5.