Add support for "modeline" in Puppet classes
A modeline allows one to add properties to a Puppet class.

A modeline is a line of text in the documentation of a class, starting
by "foreman:" and ending at the first newline character. "foreman:" may
only be preceded by whitespaces.

The smart-proxy doesn't know the format of this modeline. The text,
without the "foreman:" prefix, is returned with the name and the
parameters of a class and can be found using the "modeline" key.

Example:
    # Class: mymodule::myclass
    #
    # foreman: { internal-class }
    class mymodule::myclass {
        ...
    }